removed pandas dependency from EikonGetTimeseries
pandas verion change from 0.25 to 1.0.0 lead to having to specify the NA type. This was not possible. In stead it was better to remove the dependency all together. Furter additional measures have been taken to increase the sturdiness of requests to prevent 400 (Bad Request) and 408 (time out request) from refinitiv