Updated GetHistory and GetData
GetHistory and GetData will no longer melt data but instead return the columns as shown in Python also date classes in Columns are respected

furthermore lots of bugfixes and updates.

GetHistory now also checks parameter fields for dates in shape of yyy-mm-dd so that proper start and end arguments are respected.

Lots of new tests are provided to check if JSON and Python connections provide identical results. So that both connection methods can be used interchangeable.