json method now can use tickets assigned by udf
in some cases a request can not be dealt with immediately.  In this case a ticket is assigned. In the past when this happened the request  was simply resend  leading to  more waiting and when it was very busy the response was never received. Now the body of the request is rewritten and instead of resending the old request a new request with the ticket is requested. This leads hopefully to better performance of JSON connection method