fix correct naming of nested list returns
When RDPsearch is used in some cases nested list are returned by the api like in the case of e.g. (RCSIssueCountryRegisteredForSale, RCSAssetCategoryGenealogy1) in the past the returned column names between python and json connection methods different or sometimes failed to provide a response at all. This updates fixes this behaviour.